Folder restructuring ☘️
