--- src/ignitus-Authentication/ignitus-StudentSignUp/Components/index.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export {default as Signup} from '../signUp/Components/signUp';
