+++ src/ignitus-Authentication/ignitus-StudentSignUpFlow/emailConfirmed/Components/emailConfirmed.js
