+++ src/ignitus-Authentication/ignitus-StudentSignUpFlow/ignitus-StudentSignUp/Components/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as Signup } from './SignUp';
