+++ src/ignitus-Authentication/ignitus-StudentSignUpFlow/ignitus-StudentSignUp/selectors.js
@@ -1,5 +1,5 @@
 import { createSelector } from 'reselect';
-import { isObject } from '../../ignitus-Shared/ignitus-Utilities/HelperFunctions/lodashHelpers';
+import { isObject } from '../../../ignitus-Shared/ignitus-Utilities/HelperFunctions/lodashHelpers';
 
 export const selectStudentState = state => state.studentSignUpReducer;
