+++ src/rootReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import testimonial from './ignitus-HomePage/ignitus-Testimonial';
-import studentSignUp from './ignitus-Authentication/ignitus-StudentSignUp';
+import studentSignUp from './ignitus-Authentication/ignitus-StudentSignUpFlow/ignitus-StudentSignUp';
 import studentLogin from './ignitus-Authentication/ignitus-StudentLogin';
 import contributors from './ignitus-AboutPage/ignitus-Team';
